refactor:Props Drilling해결
